requirements_dev.txt
certifi==2017.11.5
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
coverage==4.4.2
defusedxml==0.5.0
dj-database-url==0.4.2
dj-static==0.0.6
Django==2.0
django-debug-toolbar==1.9.1
django-rest-swagger==2.1.2
django-templated-mail==1.0.0
django-toolbelt==0.0.1
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
djoser==1.1.5
gunicorn==19.7.1
idna==2.6
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.0
oauthlib==2.0.6
openapi-codec==1.3.2
psycopg2==2.7.3.2
PyJWT==1.5.3
python3-openid==3.1.0
pytz==2017.3
requests==2.18.4
requests-oauthlib==0.8.0
simplejson==3.13.2
six==1.11.0
sqlparse==0.2.4
static3==0.7.0
uritemplate==3.0.0
urllib3==1.22
